# PERF: sparse take