--- a/pandas/core/arrays/sparse/array.py
+++ b/pandas/core/arrays/sparse/array.py
@@ -1042,7 +1042,9 @@ class SparseArray(OpsMixin, PandasObject, ExtensionArray):
         sp_indexer = self.sp_index.lookup_array(indices)
         value_mask = sp_indexer != -1
         new_sp_values = self.sp_values[sp_indexer[value_mask]]
-        value_indices = np.flatnonzero(value_mask).astype(np.int32)
+
+        value_indices = np.flatnonzero(value_mask).astype(np.int32, copy=False)
+
         new_sp_index = make_sparse_index(len(indices), value_indices, kind=self.kind)
         return type(self)._simple_new(new_sp_values, new_sp_index, dtype=self.dtype)